chore: simplify bundling
