+++ .prettierignore
@@ -16,5 +16,4 @@ npm-debug.log
 
 index.cjs
 src/postcss-calc-ast-parser.ts
-src/deepmerge.ts
 dist/
