+++ package.json
@@ -21,11 +21,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "rimraf dist && run-p esm:* && run-p build:*",
-    "build:cjs": "rollup -c rollup/cjs.config.mjs",
-    "build:main": "tsc -p tsconfig.build.json",
-    "esm:deepmerge": "rollup -c rollup/deepmerge.config.mjs",
-    "esm:postcss-calc": "rollup -c rollup/postcss-calc.config.mjs",
+    "build": "rimraf dist && rollup -c rollup.config.mjs",
     "format": "npm run format:eslint && npm run format:prettier",
     "format:eslint": "eslint --ext .ts,.html . --fix",
     "format:prettier": "prettier \"**/*.{ts,md}\" \"package.json\" --write",
@@ -44,6 +40,7 @@
   "types": "./dist/index.d.ts",
   "dependencies": {
     "@tokens-studio/types": "^0.2.1",
+    "browser-style-dictionary": "^3.1.1-browser.1",
     "color2k": "^2.0.1",
     "colorjs.io": "^0.4.3",
     "deepmerge": "^4.3.1",
@@ -61,6 +58,7 @@
     "@typescript-eslint/eslint-plugin": "^5.54.0",
     "@typescript-eslint/parser": "^5.54.0",
     "@web/dev-server-esbuild": "^0.3.3",
+    "@web/dev-server-rollup": "^0.5.0",
     "@web/test-runner": "^0.15.1",
     "@web/test-runner-playwright": "^0.9.0",
     "eslint": "^8.32.0",
