+++ web-test-runner.config.mjs
@@ -1,5 +1,9 @@
 import { playwrightLauncher } from '@web/test-runner-playwright';
 import { esbuildPlugin } from '@web/dev-server-esbuild';
+import { fromRollup } from '@web/dev-server-rollup';
+import commonjsRollup from '@rollup/plugin-commonjs';
+
+const commonjs = fromRollup(commonjsRollup);
 
 export default {
   nodeResolve: true,
@@ -13,8 +17,8 @@ export default {
       functions: 100,
       lines: 100,
     },
-    exclude: ['src/postcss-calc-ast-parser.ts', 'src/deepmerge.ts', 'node_modules/**'],
+    exclude: ['src/postcss-calc-ast-parser.ts', 'node_modules/**'],
   },
   browsers: [playwrightLauncher({ product: 'chromium' })],
-  plugins: [esbuildPlugin({ ts: true, target: 'auto' })],
+  plugins: [commonjs({ requireReturnsDefault: true }), esbuildPlugin({ ts: true, target: 'auto' })],
 };
